+++ src/SMAPI/Framework/ContentManagers/BaseContentManager.cs
@@ -138,6 +138,9 @@ namespace StardewModdingAPI.Framework.ContentManagers
             }
         }
 
+        /// <summary>Perform any cleanup needed when the locale changes.</summary>
+        public virtual void OnLocaleChanged() { }
+
         /// <summary>Normalise path separators in a file path. For asset keys, see <see cref="AssertAndNormaliseAssetName"/> instead.</summary>
         /// <param name="path">The file path to normalise.</param>
         [Pure]